+++ js/welcome.js
@@ -1,7 +1,7 @@
 import creatDOMElement from '../js/create-dom-element.js';
 import showPage from '../js/show-page.js';
 import {levelArtist} from '../js/level-artist.js';
-import {page} from '../js/game.js';
+import {showRandomPage} from '../js/show-random-page.js';
 
 /* 1 отображение страницы с артистом при клике на плей */
 const welcomeMarkup = `
@@ -22,7 +22,7 @@ const welcome = {
   page: () => creatDOMElement(welcomeMarkup, welcomeClassName),
   init: () => {
     const onMainPlayClick = () => {
-      page.number++;
+      showRandomPage();
       showPage(levelArtist);
     };
